+++ src/main/scala/scorex/waves/transaction/WavesTransactionModule.scala
@@ -1,12 +1,15 @@
 package scorex.waves.transaction
 
 import scorex.account.{Account, PrivateKeyAccount, PublicKeyAccount}
+import scorex.api.http.NoBalance
 import scorex.app.Application
 import scorex.block.BlockField
 import scorex.crypto.encode.Base58
 import scorex.settings.Settings
 import scorex.transaction._
 import scorex.utils.NTP
+
+import scala.util.{Failure, Success, Try}
 
 /**
   * Waves Transaction Module
@@ -24,16 +27,19 @@ class WavesTransactionModule(implicit override val settings: TransactionSettings
   /**
     * Publish signed payment transaction which generated outside node
     */
-  def broadcastPayment(externalPayment: ExternalPayment): PaymentTransaction = {
+  def broadcastPayment(externalPayment: ExternalPayment): Try[PaymentTransaction] = {
     val time = externalPayment.timestamp
     val sigBytes = Base58.decode(externalPayment.signature).get
     val senderPubKey = Base58.decode(externalPayment.senderPublicKey).get
     val recipientAccount = new Account(externalPayment.recipient)
     val payment = new PaymentTransaction(new PublicKeyAccount(senderPubKey),
       recipientAccount, externalPayment.amount, externalPayment.fee, time, sigBytes)
-    if (blockStorage.state.isValid(payment))
+    if (blockStorage.state.isValid(payment)) {
       onNewOffchainTransaction(payment)
-    payment
+      Success(payment)
+    } else {
+      Failure(new Exception(NoBalance.message))
+    }
   }
 
   override def genesisData: BlockField[SimpleTransactionModule.StoredInBlock] = {